+++ ui/src/App.tsx
@@ -9,6 +9,7 @@ import { Toaster } from "@/components/ui/sonner";
 import { LoginPage } from "@/pages/login/LoginPage";
 import { DashboardPage } from "@/pages/dashboard/DashboardPage";
 import { RepositoriesPage } from "@/pages/repositories/RepositoriesPage";
+import { RepositoryDetailPage } from "@/pages/repositories/RepositoryDetailPage";
 import { ContentViewsPage } from "@/pages/content-views/ContentViewsPage";
 import { EnvironmentsPage } from "@/pages/environments/EnvironmentsPage";
 import { ServersPage } from "@/pages/servers/ServersPage";
@@ -55,6 +56,7 @@ export default function App() {
             <Route path="/login" element={<LoginPage />} />
             <Route path="/" element={<Shell><DashboardPage /></Shell>} />
             <Route path="/repositories" element={<Shell><RepositoriesPage /></Shell>} />
+            <Route path="/repositories/:name" element={<Shell><RepositoryDetailPage /></Shell>} />
             <Route path="/content-views" element={<Shell><ContentViewsPage /></Shell>} />
             <Route path="/environments" element={<Shell><EnvironmentsPage /></Shell>} />
             <Route path="/servers" element={<Shell><ServersPage /></Shell>} />